up to 89 pct coverage
