+++ aglio/sample_data/wUS-SH-2010_TNA_ak135_ref.csv
@@ -1,0 +1,141 @@
+depth_km,Vs,Vp
+38.0,4.8,8.601422479005231
+50.0,4.79,8.585819405882608
+75.0,4.775,8.567671820094573
+100.0,4.775,8.549063223830736
+125.0,4.775,8.555229866193539
+150.0,4.775,8.619004197161233
+175.0,4.71,8.564591151050905
+200.0,4.63,8.480976582288811
+225.0,4.64,8.52841596073045
+250.0,4.67,8.590813226712342
+275.0,4.695,8.646529638869447
+300.0,4.72,8.703979309050132
+325.0,4.74,8.752117899318256
+350.0,4.755,8.79097398015531
+375.0,4.765,8.820296772796818
+405.0,4.78,8.860826213232086
+406.0,5.0,9.269091699804395
+460.0,5.1864,9.528
+510.0,5.2922,9.6962
+560.0,5.3989,9.864
+610.0,5.5047,10.032
+660.0,5.6104,10.2
+660.0,5.9607,10.7909
+710.0,6.0898,10.9222
+760.0,6.21,11.0553
+809.5,6.2424,11.1355
+859.0,6.2799,11.2228
+908.5,6.3164,11.3068
+958.0,6.3519,11.3897
+1007.5,6.386,11.4704
+1057.0,6.4182,11.5493
+1106.5,6.4514,11.6265
+1156.0,6.4822,11.702
+1205.5,6.5131,11.7768
+1255.0,6.5431,11.8491
+1304.5,6.5728,11.9208
+1354.0,6.6009,11.9891
+1403.5,6.6285,12.0571
+1453.0,6.6554,12.1247
+1502.5,6.6813,12.1912
+1552.0,6.707,12.2558
+1601.5,6.7323,12.3181
+1651.0,6.7579,12.3813
+1700.5,6.782,12.4427
+1750.0,6.8056,12.503
+1799.5,6.8289,12.5638
+1849.0,6.8517,12.6226
+1898.5,6.8743,12.6807
+1948.0,6.8972,12.7384
+1997.5,6.9194,12.7956
+2047.0,6.9416,12.8524
+2096.5,6.9625,12.9093
+2146.0,6.9852,12.9663
+2195.5,7.0069,13.0226
+2245.0,7.0286,13.0786
+2294.5,7.0504,13.1337
+2344.0,7.0722,13.1895
+2393.5,7.0932,13.2465
+2443.0,7.1144,13.3017
+2492.5,7.1368,13.3584
+2542.0,7.1584,13.4156
+2591.5,7.1804,13.4741
+2640.0,7.2031,13.5311
+2690.0,7.2253,13.5899
+2740.0,7.2485,13.6498
+2740.0,7.2485,13.6498
+2789.67,7.2593,13.6533
+2839.33,7.27,13.657
+2891.5,7.2817,13.6601
+2891.5,0.0,8.0
+2939.33,0.0,8.0382
+2989.66,0.0,8.1283
+3039.99,0.0,8.2213
+3090.32,0.0,8.3122
+3140.66,0.0,8.4001
+3190.99,0.0,8.4861
+3241.32,0.0,8.5692
+3291.65,0.0,8.6496
+3341.98,0.0,8.7283
+3392.31,0.0,8.8036
+3442.64,0.0,8.8761
+3492.97,0.0,8.9461
+3543.3,0.0,9.0138
+3593.64,0.0,9.0792
+3643.97,0.0,9.1426
+3694.3,0.0,9.2042
+3744.63,0.0,9.2634
+3794.96,0.0,9.3205
+3845.29,0.0,9.376
+3895.62,0.0,9.4297
+3945.95,0.0,9.4814
+3996.28,0.0,9.5306
+4046.62,0.0,9.5777
+4096.95,0.0,9.6232
+4147.28,0.0,9.6673
+4197.61,0.0,9.71
+4247.94,0.0,9.7513
+4298.27,0.0,9.7914
+4348.6,0.0,9.8304
+4398.93,0.0,9.8682
+4449.26,0.0,9.9051
+4499.6,0.0,9.941
+4549.93,0.0,9.9761
+4600.26,0.0,10.0103
+4650.59,0.0,10.0439
+4700.92,0.0,10.0768
+4751.25,0.0,10.1095
+4801.58,0.0,10.1415
+4851.91,0.0,10.1739
+4902.24,0.0,10.2049
+4952.58,0.0,10.2329
+5002.91,0.0,10.2565
+5053.24,0.0,10.2745
+5103.57,0.0,10.2854
+5153.5,0.0,10.289
+5153.5,3.5043,11.0427
+5204.61,3.5187,11.0585
+5255.32,3.5314,11.0718
+5306.04,3.5435,11.085
+5356.75,3.5551,11.0983
+5407.46,3.5661,11.1166
+5458.17,3.5765,11.1316
+5508.89,3.5864,11.1457
+5559.6,3.5957,11.159
+5610.31,3.6044,11.1715
+5661.02,3.6126,11.1832
+5711.74,3.6202,11.1941
+5762.45,3.6272,11.2041
+5813.16,3.6337,11.2134
+5863.87,3.6396,11.2219
+5914.59,3.645,11.2295
+5965.3,3.6498,11.2364
+6016.01,3.654,11.2424
+6066.72,3.6577,11.2477
+6117.44,3.6608,11.2521
+6168.15,3.6633,11.2557
+6218.86,3.6653,11.2586
+6269.57,3.6667,11.2606
+6320.29,3.6675,11.2618
+6371.0,3.6678,11.2622
